+++ pyflp/flobject/arrangement/timemarker.py
@@ -16,7 +16,7 @@ class TimeMarker(FLObject):
 
     @name.setter
     def name(self, value: str):
-        self.setprop("name", value)
+        self._setprop("name", value)
 
     @property
     def position(self) -> Optional[int]:
@@ -26,7 +26,7 @@ class TimeMarker(FLObject):
 
     @position.setter
     def position(self, value: int):
-        self.setprop("position", value)
+        self._setprop("position", value)
 
     @property
     def numerator(self) -> Optional[int]:
@@ -36,7 +36,7 @@ class TimeMarker(FLObject):
     @numerator.setter
     def numerator(self, value: int):
         assert value in range(1, 17)
-        self.setprop("numerator", value)
+        self._setprop("numerator", value)
 
     @property
     def denominator(self) -> Optional[int]:
@@ -46,22 +46,22 @@ class TimeMarker(FLObject):
     @denominator.setter
     def denominator(self, value: int):
         assert value in (2, 4, 8, 16)
-        self.setprop("denominator", value)
+        self._setprop("denominator", value)
 
     # * Parsing logic
     def _parse_byte_event(self, e: ByteEvent):
         if e.id == TimeMarkerEvent.Numerator:
-            self.parse_uint8_prop(e, "numerator")
+            self._parse_uint8_prop(e, "numerator")
         elif e.id == TimeMarkerEvent.Denominator:
-            self.parse_uint8_prop(e, "denominator")
+            self._parse_uint8_prop(e, "denominator")
 
     def _parse_dword_event(self, e: DWordEvent):
         if e.id == TimeMarkerEvent.Position:
-            self.parse_uint32_prop(e, "position")
+            self._parse_uint32_prop(e, "position")
 
     def _parse_text_event(self, e: TextEvent):
         if e.id == TimeMarkerEvent.Name:
-            self.parse_str_prop(e, "name")
+            self._parse_str_prop(e, "name")
 
     def __init__(self):
         super().__init__()